Use a PetitionProject Object
