+++ config/settings/base.py
@@ -83,6 +83,7 @@ LOCAL_APPS = [
     "envergo.admin.config.EnvergoAdminConfig",
     "envergo.urlmappings",
     "envergo.hedges",
+    "envergo.petitions",
 ]
 # https://docs.djangoproject.com/en/dev/ref/settings/#installed-apps
 INSTALLED_APPS = DJANGO_APPS + THIRD_PARTY_APPS + LOCAL_APPS
