+++ envergo/pages/urls_haie.py
@@ -2,9 +2,7 @@ from django.urls import path
 from django.utils.translation import gettext_lazy as _
 from django.views.generic import RedirectView, TemplateView
 
-from envergo.moulinette.views import PetitionProjectDetail
 from envergo.pages.views import (
-    DemarcheSimplifieeView,
     HomeHaieView,
     LegalMentionsView,
     Outlinks,
@@ -33,14 +31,4 @@ urlpatterns = [
         name="contact_us",
     ),
     path("admin/outlinks/", Outlinks.as_view(), name="outlinks"),
-    path(
-        "demarchesimplifiee/",
-        DemarcheSimplifieeView.as_view(),
-        name="demarchesimplifiee",
-    ),
-    path(
-        "projet/",
-        PetitionProjectDetail.as_view(),
-        name="projet",
-    ),
 ]
