+++ envergo/pages/views.py
@@ -6,7 +6,7 @@ import requests
 from django.conf import settings
 from django.contrib import messages
 from django.contrib.syndication.views import Feed
-from django.http import HttpResponseRedirect, JsonResponse
+from django.http import HttpResponseRedirect
 from django.urls import reverse
 from django.utils.formats import date_format
 from django.utils.html import mark_safe
@@ -16,9 +16,7 @@ from config.settings.base import GEOMETRICIAN_WEBINAR_FORM_URL
 from envergo.geodata.models import Department
 from envergo.moulinette.models import ConfigAmenagement
 from envergo.moulinette.views import MoulinetteMixin
-from envergo.pages.forms import DemarcheSimplifieeForm
 from envergo.pages.models import NewsItem
-from envergo.urlmappings.models import UrlMapping
 
 logger = logging.getLogger(__name__)
 
@@ -278,69 +276,3 @@ class NewsFeed(Feed):
         base_url = reverse("faq_news")
         item_url = f"{base_url}#news-item-{item.id}"
         return item_url
-
-
-class DemarcheSimplifieeView(FormView):
-    form_class = DemarcheSimplifieeForm
-
-    def form_valid(self, form):
-        moulinette_url = form.cleaned_data["moulinette_url"]
-        profil = form.cleaned_data["profil"]
-
-        read_only_mapping = UrlMapping.objects.create(url=moulinette_url)
-        read_only_url = reverse(
-            "moulinette_result_read_only",
-            kwargs={"moulinette_reference": read_only_mapping.key},
-        )
-        # Ce code est particulièrement fragile.
-        # Un changement dans un label côté démarche simplifiées cassera ce mapping sans prévenir.
-        mapping_demarche_simplifiee = {
-            "autre": "Autre (collectivité, aménageur, gestionnaire de réseau, particulier, etc.)",
-            "agri_pac": "Exploitant-e agricole bénéficiaire de la PAC",
-        }
-        demarche_id = settings.DEMARCHES_SIMPLIFIEE["DEMARCHE_HAIE"]["ID"]
-        api_url = f"{settings.DEMARCHES_SIMPLIFIEE['API_URL']}demarches/{demarche_id}/dossiers"
-
-        body = {
-            settings.DEMARCHES_SIMPLIFIEE["DEMARCHE_HAIE"][
-                "PROFIL_FIELD_ID"
-            ]: mapping_demarche_simplifiee[profil],
-            settings.DEMARCHES_SIMPLIFIEE["DEMARCHE_HAIE"][
-                "MOULINETTE_URL_FIELD_ID"
-            ]: moulinette_url,
-        }
-
-        response = requests.post(
-            api_url, json=body, headers={"Content-Type": "application/json"}
-        )
-        demarche_simplifiee_url = None
-
-        if 200 <= response.status_code < 400:
-            data = response.json()
-            demarche_simplifiee_url = data.get("dossier_url")
-        else:
-            logger.error(
-                "Error while pre-filling a dossier on demarches-simplifiees.fr",
-                extra={"response": response},
-            )
-
-        if not demarche_simplifiee_url:
-            res = self.form_invalid(form)
-        else:
-            res = JsonResponse(
-                {
-                    "demarche_simplifiee_url": demarche_simplifiee_url,
-                    "read_only_url": read_only_url,
-                }
-            )
-
-        return res
-
-    def form_invalid(self, form):
-        return JsonResponse(
-            {
-                "error_title": "Un problème technique empêche la création de votre dossier.",
-                "error_body": "Nous vous invitons à enregistrer votre simulation et à réessayer ultérieurement.",
-            },
-            status=400,
-        )
