+++ envergo/petitions/models.py
@@ -1,0 +1,31 @@
+from django.db import models
+from django.utils import timezone
+from django.utils.translation import gettext_lazy as _
+
+from envergo.evaluations.models import generate_reference
+
+
+class PetitionProject(models.Model):
+    """A petition project by a project owner."""
+
+    reference = models.CharField(
+        _("Reference"),
+        max_length=64,
+        null=True,
+        default=generate_reference,
+        unique=True,
+        db_index=True,
+    )
+    moulinette_url = models.URLField(_("Moulinette url"), max_length=1024, blank=True)
+
+    hedges_data = models.JSONField()
+
+    # Meta fields
+    created_at = models.DateTimeField(_("Date created"), default=timezone.now)
+
+    class Meta:
+        verbose_name = _("Petition project")
+        verbose_name_plural = _("Petition projects")
+
+    def __str__(self):
+        return self.reference
